requirements.txt
fastapi
uvicorn
pandas
numpy
ta
httpx
python-dotenv
